Replace `player.GetAll` calls with a PVS player cache
Constantly calling `player.GetAll` can greatly reduce performance (due to C->Lua calls), the solution to this problem is to use some kind of cache that will retrieve the players that will be visible to the player. This way the performance will only be done with the cached players instead of asking for a copy of the players on the engine side. This method does not take away from the robustness of the script because the `NotifyShouldTransmit` event has no problems so far to signal the change of players' state.